Hadrien: ajout Client + Commande + LigneCommande

+++ Ecommerce_Spring/Projet_Spring_Ecommerce/src/main/java/fr/adaming/model/LigneCommande.java
@@ -36,12 +36,12 @@ public class LigneCommande implements Serializable {
 	public LigneCommande() {
 		super();
 	}
-	public LigneCommande(int quantite, int prix) {
+	public LigneCommande(int quantite, double prix) {
 		super();
 		this.quantite = quantite;
 		this.prix = prix;
 	}
-	public LigneCommande(long id, int quantite, int prix) {
+	public LigneCommande(long id, int quantite, double prix) {
 		super();
 		this.id = id;
 		this.quantite = quantite;
